1.0.9 - start using variables and schemas
We leverage local variables and global (in "config") schemas to
start vetting thesse features, fixing issues along the way.

One issue noticed in general was that a common dependency (i.e. a
"compile") would get resolved, a single consumer would get updated
but the "is this target satisfied" logic would then short-circuit
the other targets, noting that the dependency (correctly) was
already resolved.

Fixes this issue by adding a new notion of "was this dependency
newly resolved".