1.0.10 - Improve file-change detection correctness
Fixes an issue where new/removed/renamed files are incorrectly
detected as not changed because the cache has a stale (but
correct) hash for that file that has never been cleared.

Forces the cache to sync the correctness of its state with
the actual file system upon load and before writing.